Update jquery.jkit.1.1.14.js
Changed how tabs are determined. Previously tabs relied on <h3>followed by <div> for tab and tab content respectively. <h3> tags are semantic tags and might not be appropriate in their location in HTML. By  changing the selectors to :first-child and :last-child selection of tabs  and tab content are agnostic of HTML and only require two elements that follow one another. This also allows being able to set much richer HTML for tabs.